operator: Refactor k8s node watcher for re-usability
This commit is to perform lift and shift current initialization for k8s
node watcher to another function with single scope of work, so that it
can be re-used later. There is no change in logic.
